requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
pymysql
cryptography
aiokafka
boto3
python-multipart
python-dotenv
pydantic[email]
passlib[bcrypt]
python-jose[cryptography]
slowapi